Implement Cash Movements functionality with repository pattern
- Added ICashMovementsRepository interface for CRUD operations on cash movements.
- Created CashMovements, CashMovementsRequest, and CashMovementsResponse models.
- Implemented CashMovementsRepository to handle database interactions.
- Updated database migrations to include CashMovements table and related changes.
- Modified ProductSales and Sales tables to accommodate new data types.
- Adjusted project configuration files for compatibility with the new structure.

+++ Models/Sale/SaleResponse.cs
@@ -10,7 +10,7 @@ namespace pruebaAPI.Models
 
         public Guid Id { set; get; }
         public DateTime SaleDate { set; get; }
-        public double Amount { set; get; }
+        public decimal Amount { set; get; }
         public int Quantity { set; get; }
     }
 }
